requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
Django==3.1
django-appconf==1.0.4
djangorestframework==3.11.1
idna==2.10
libsass==0.20.1
pkg-resources==0.0.0
psycopg2==2.8.5
pytz==2020.1
rcssmin==1.0.6
requests==2.24.0
rjsmin==1.1.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
